A pretty sick performance hack.
The idea is the following: instead of calculating the LDA functionals
everytime, one can "spline" them at the beginning, and then interpolate
every time, since they are very smooth functions. Some comments:

(*) The calculation of the LDA exchange and correlation functionals do
not take a significant calculation time, *except* if one wants to use the
SIC+LDA within the OEP/KLI approximation for a big molecule.
(*) This commit is still experimental, and in fact the code is still
exactly the same as it was, if you do not compile adding the LDA_SPEEDUP
preprocessor variable (by doing CFLAGS="-O3 -DLDA_SPEEDUP", for example).
So you do not have to worry at all about this for the moment...
I will make sure that it works well and that it makes sense before making
it definive.


git-svn-id: http://www.tddft.org/svn/octopus/trunk/libxc@1061 53d8e6fd-f412-0410-a60b-8e56063eb7bf