Bug fixed.
The numerical functional derivation was not working properly in some
spin-polarized cases (Perdew-Zunger parameterization, Na4(+1), it was
just giving lots of NaN).  Now, at least for those cases, it seems to
work. I still have to do some more testing on the Casida module,
though.

In any case, the PZ kernel should be obtained analytically. Maybe in the
future.


git-svn-id: http://www.tddft.org/svn/octopus/trunk/libxc@1641 53d8e6fd-f412-0410-a60b-8e56063eb7bf