Sorry, there are several things in this commit:
*) I finally understood the problem that Alberto was having. The 
variable i in xc_functl_write_info was used for two things. In the first 
it should be a normal integer, and in the second it should be a pointer. 
That is why the code was crashing. I simply solved the problem by using 
two variables with different names.

*) Added a macro called C_POINTER that has the obvious meaning of 
integer(POINTER_SIZE). I also changed all occurences in the code of 
integer(POINTER_SIZE) to C_POINTER. I believe it is much more readable 
like that.

*) Added two new functionals, Perdew Wang 86 and the exchange part of 
Perdew Wang 91. I think the introduction of new functionals will slow 
down now ;)


git-svn-id: http://www.tddft.org/svn/octopus/trunk/libxc@2470 53d8e6fd-f412-0410-a60b-8e56063eb7bf