I am resyncing my tree that started diverging ;)
*) Further cleanup of libxc. Now it was fxc that now has only 3 
components (uu, ud, dd) for spin-polarized calculations. This amounted 
to changes in libxc and in octopus. Note that kxc is still not changed. 
It should include only the independent elements! Maybe Xavier would like 
to do it ;))

*) Added a Perdew Wang modified lda functional. This is the functional 
used by Kieron in his implementation of the PBE (that i assume is the 
reference), and that is different from the published functional 
(different number of digits in some costants, and he even has 
incompatible definitions for the same constants). In this way, the PBE 
now gives the same as the reference numbers, so I rehabilitated the 
test. The differences to the previous implementation are very small (4th 
or 5th digit, but anyway...)

*) Now we can have all-electron species also in 1 and 2 dimensions.




git-svn-id: http://www.tddft.org/svn/octopus/trunk/libxc@2511 53d8e6fd-f412-0410-a60b-8e56063eb7bf