Optimization:
* Wrote a C version of dnl_operator_operate. The inner loop is
unrolled by hand to depth 3 and prefetching directives are used to
keep the array of weights in the cache.

For a test case (benzene gs with non interacting electrons, h=2.0
r=6.0), total performance was increased by a factor of 1.47 for ifort
9.1 and by a factor of 2.2 for gfortran (p4 3.0GHz). In both cases
gcc-4.1 was used.

The new function is not enabled by default yet, but it is selected by
the input variable OperateFunction, values can be either 'fortran' or
'c'.

* Configure detects the presence of builtin functions __builtin_expect
and __builtin_prefetch in the C compiler. If they are not found,
alternative dummy macros are provided.

Fixed a couple of bugs revealed by gcc-2.95:

* Assert is not always available in C. Put a check in configure.ac.
* Move a declaration to the beginning of the function.



git-svn-id: http://www.tddft.org/svn/octopus/trunk/libxc@2590 53d8e6fd-f412-0410-a60b-8e56063eb7bf