This is a boring maintenance commit concerning the automake build and distribution system. It is likely that some of the nightly builds will fail or that someone encounters building problems. In that case, just bug me, I feel in charge to fix that.
I wanted to get VPATH builds working, i. e. compiling in a directory that
does not contain the source code. This makes it easily possible to
configure and build a single source tree with lots of different options
(avoiding tedious copying). VPATH is basically a GNU make feature but
the Makfefile.ams have to follow certain restrictions to get it working:

1. Relative file names should be constructed with the $(srcdir),
$(top_srcdir), $(builddir), $(top_builddir) variables.

2. Wildcards should not be used in automake variables. This means
that all source files have to be mentioned in the Makefile.am. This
is a bit tiresome (especially in the PP directory) but it is
recommended in the GNU Coding Standard and considered good practice.
So, please add any .inp and .test files you include in the testsuite
to the corresponding Makefile.ams. This also holds for the .psf files
(and the .hgh) if a new element should be discovered...

3. I had to change the way the variable documentation is generated a bit.
The main thing was to add command-line flags to specify input and output
directories to the var2html.pl and mk_varinfo.pl scripts.

4. Filename flags for the compilers (like -I) should not be added to the
corresponding variable in the configure.ac but to the AM_ version in
the approripate Makefile.am (or to its per target variant). This
makes use of $(srcdir), etc. possible. In some Makefile.ams the
compilation rule is redefined, in these cases, it has to be checked if
the AM_ variant is included (e. g. src/Makefile.am).

There were some more things I ran across:

1. The testsuite can only be run after a make install because
the share data is not available otherwise. For that reason, I replaced
the check target by an installcheck-local rule.

2. The oct-run_testsuite script used hardcoded paths, they are now
patched by the configure script. So, the oct-run_testuite is replaced
by oct-run_testsuite.in and the original script will disappear in my
next commit.

3. The libxc testuite also needed some fixing to consider different
source and build trees.

4. I completely removed the AC_CONFIG_SUBDIRS for the python tree.
The conditional configuration of a subtree is a very subtle issue as
automake does not support not configured subtrees. The configure.ac and
Makefile.am of the python part seem to be in a very fragile state, and as
it is not used anyway at the moment, I did not spend time to fix them.
By the way, how is the state of the GUI? There is just few lines of code
in the repository.
When it will be included again, I recommend just another Makefile.am
and no subpackage configure. Necessary checks should be added to our
main configure.ac (and can be skipped, if no GUI is requested).

The code now passes the distcheck target which is desireable.


git-svn-id: http://www.tddft.org/svn/octopus/trunk/libxc@2771 53d8e6fd-f412-0410-a60b-8e56063eb7bf