Moved the fortran interface of libxc to the libxc/src/ directory. This
Instead of the C_POINTER, I defined some derived types (with a pointer
inside), so in theory they should have enough space to hold a pointer
but I have to test it. This approach is also more elegant as the types
can't be passed incorrectly and their contents are private, so the
user can't modify them without libxc.



git-svn-id: http://www.tddft.org/svn/octopus/trunk/libxc@2995 53d8e6fd-f412-0410-a60b-8e56063eb7bf