Bug correction from Jussi Enkovaara:
I found a small bug in the PBE of libxc. In the function "perdew_params" 
in the file "gga_perdew.c" the total gradient "pt->gdmt"

   /* get gdmt = |nabla n| */
   pt->gdmt = sigma[0];
   if(pt->nspin == XC_POLARIZED) pt->gdmt += 2.0*sigma[1] + sigma[2];

can become a small negative number (in my antiferromagnetic system ~ -1e-20), and
the following square root produces then NaN. I fixed that with a simple if
statement i.e.

   /* get gdmt = |nabla n| */
   pt->gdmt = sigma[0];
   if(pt->nspin == XC_POLARIZED) pt->gdmt += 2.0*sigma[1] + sigma[2];
   if(pt->gdmt < 0.0) pt->gdmt = MIN_GRAD*MIN_GRAD;
   pt->gdmt = sqrt(pt->gdmt);



git-svn-id: http://www.tddft.org/svn/octopus/trunk/libxc@3667 53d8e6fd-f412-0410-a60b-8e56063eb7bf