Adding KS inversion as a new family of functionals in a spirit similar to OEP. There is now a new module xc_ks_inversion_m that contains most of the inversion code. I have adapted the 'old' runmode invert_ks to use this new infrastructure.
To deal with dependencies I had to move the eigensolver
module from scf to system (as suggested by Miguel). This
is analogous to the linear solver in the case of OEP.

Not all things are in place yet to use the functional
family in the general case, however the old runmode
should work as before.


git-svn-id: http://www.tddft.org/svn/octopus/trunk/libxc@6389 53d8e6fd-f412-0410-a60b-8e56063eb7bf